homebrew: Replacement for `brew bundle`
Apparently `brew bundle` is deprecated and might, maybe be replaced in
the future. Until then we’ll run with this little thingy.

Disable Caskfile in `post-up` hook because I have quite a lot of apps
that haven’t been installed through `cask`, and I really can’t bother
with it right now.